+++ src/main/java/org/owasp/webgoat/lessons/webwolfintroduction/MailAssignment.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.webwolf_introduction;
+package org.owasp.webgoat.lessons.webwolfintroduction;
 
 import org.apache.commons.lang3.StringUtils;
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
