+++ src/main/resources/lessons/webwolfintroduction/html/WebWolfIntroduction.html
@@ -2,15 +2,15 @@
 <html xmlns:th="http://www.thymeleaf.org">
 
 <div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:lessons/webwolf_introduction/documentation/IntroductionWebWolf.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/webwolfintroduction/documentation/IntroductionWebWolf.adoc"></div>
 </div>
 
 <div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:lessons/webwolf_introduction/documentation/Uploading_files.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/webwolfintroduction/documentation/Uploading_files.adoc"></div>
 </div>
 
 <div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:lessons/webwolf_introduction/documentation/Receiving_mail.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/webwolfintroduction/documentation/Receiving_mail.adoc"></div>
     <div class="attack-container">
         <img th:src="@{/images/wolf-enabled.png}" class="webwolf-enabled"/>
         <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
@@ -66,7 +66,7 @@
 </div>
 
 <div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:lessons/webwolf_introduction/documentation/Landing_page.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/webwolfintroduction/documentation/Landing_page.adoc"></div>
     <div class="attack-container">
         <img th:src="@{/images/wolf-enabled.png}" class="webwolf-enabled"/>
         <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
