+++ src/main/resources/lessons/webwolfintroduction/i18n/WebGoatLabels.properties
