+++ src/main/resources/lessons/webwolfintroduction/images/wolf-enabled.png
